+++ dorm/__init__.py
@@ -31,6 +31,7 @@ Quick start:
 
 from .conf import configure, settings
 from . import signals, transaction
+from .db.connection import ahealth_check, health_check
 from .signals import (
     Signal,
     post_delete,
@@ -123,6 +124,8 @@ __all__ = [
     "settings",
     "transaction",
     "signals",
+    "health_check",
+    "ahealth_check",
     # Signals
     "Signal",
     "pre_save",